Got fetchUnis() to work!
A request to a certain page sometimes partially fails,
in that it returns the wrong page or unis on other pages,
meaning I'm not certain that I have all the unis.
(There are overlaps between students and faculty/staff,
but these are real overlaps, and should be kept).

To solve this, I wrote OkHttpClient.readPaginatedRetrying(),
which, given a find num pages and num items function,
keeps sequentially requesting all the pages
until all of the items are found.

I have a bunch of debug code lying around.
I wanted to commit them and then clean them up in case I need them again.